# swagger-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/institution": {
      "post": {
        "description": "创建新的机构",
        "operationId": "InstitutionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstitutionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建机构",
        "tags": [
          "机构管理"
        ]
      },
      "get": {
        "description": "获取所有机构",
        "operationId": "InstitutionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有机构",
        "tags": [
          "机构管理"
        ]
      }
    },
    "/api/v1/institution/{id}": {
      "get": {
        "description": "获取指定机构",
        "operationId": "InstitutionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取指定机构",
        "tags": [
          "机构管理"
        ]
      },
      "put": {
        "description": "更新指定机构",
        "operationId": "InstitutionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstitutionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新机构",
        "tags": [
          "机构管理"
        ]
      },
      "delete": {
        "description": "删除指定机构",
        "operationId": "InstitutionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除机构",
        "tags": [
          "机构管理"
        ]
      }
    },
...
    "/api/v1/profile": {
      "post": {
        "description": "创建新的个人资料",
        "operationId": "ProfileController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建个人资料",
        "tags": [
          "个人资料管理"
        ]
      },
      "get": {
        "description": "获取所有个人资料",
        "operationId": "ProfileController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有个人资料",
        "tags": [
          "个人资料管理"
        ]
      }
    },
    "/api/v1/profile/{id}": {
      "get": {
        "description": "获取指定个人资料",
        "operationId": "ProfileController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取指定个人资料",
        "tags": [
          "个人资料管理"
        ]
      },
      "put": {
        "description": "更新指定个人资料",
        "operationId": "ProfileController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新个人资料",
        "tags": [
          "个人资料管理"
        ]
      },
      "delete": {
        "description": "删除指定个人资料",
        "operationId": "ProfileController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除个人资料",
        "tags": [
          "个人资料管理"
        ]
      }
    },
    "/api/v1/role/create": {
      "post": {
        "description": "创建新的角色",
        "operationId": "RoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "角色创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "summary": "创建角色",
        "tags": [
          "角色管理"
        ]
      }
    },
...
    "/api/v1/role/findOne/{id}": {
      "get": {
        "description": "根据ID获取特定角色信息",
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取角色信息成功"
          },
          "404": {
            "description": "角色不存在"
          }
        },
        "summary": "获取指定角色",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/v1/role/update/{id}": {
      "put": {
        "description": "更新指定角色的信息",
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "角色更新成功"
          },
          "404": {
            "description": "角色不存在"
          }
        },
        "summary": "更新角色",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/v1/role/delete/{id}": {
      "delete": {
        "description": "删除指定的角色",
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色删除成功"
          },
          "404": {
            "description": "角色不存在"
          }
        },
        "summary": "删除角色",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/v1/menus/create": {
      "post": {
        "description": "创建新的菜单",
        "operationId": "MenusController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "菜单创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建菜单",
        "tags": [
          "菜单管理"
        ]
      }
    },
    "/api/v1/menus/findAll": {
      "get": {
        "description": "获取所有菜单",
        "operationId": "MenusController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取菜单列表成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有菜单",
        "tags": [
          "菜单管理"
        ]
      }
    },
    "/api/v1/menus/findOne/{id}": {
      "get": {
        "description": "获取指定菜单",
        "operationId": "MenusController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取菜单信息成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取指定菜单",
        "tags": [
          "菜单管理"
        ]
      }
    },
    "/api/v1/menus/update/{id}": {
      "patch": {
        "description": "更新指定菜单",
        "operationId": "MenusController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "菜单更新成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新菜单",
        "tags": [
          "菜单管理"
        ]
      }
    },
    "/api/v1/menus/delete/{id}": {
      "delete": {
        "description": "删除指定菜单",
        "operationId": "MenusController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "菜单删除成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除菜单",
        "tags": [
          "菜单管理"
        ]
      }
    }
  },
  "info": {
    "title": "nestjs Demo",
    "description": "nestjs 项目",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "开发环境"
    },
    {
      "url": "http://127.0.0.1:3001/api/v1",
      "description": "API v1 版本"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateInstitutionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateInstitutionDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateProfileDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "角色名称",
            "example": "admin"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "example": "系统管理员"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "角色名称",
            "example": "admin"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "example": "系统管理员"
          }
        }
      },
      "CreateMenuDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "菜单名称",
            "example": "用户管理"
          },
          "path": {
            "type": "string",
            "description": "菜单路径",
            "example": "/users"
          },
          "acl": {
            "type": "string",
            "description": "权限策略",
            "example": "user:read,user:write"
          },
          "roleId": {
            "type": "number",
            "description": "角色ID",
            "example": 1
          }
        },
        "required": [
          "name",
          "path",
          "roleId"
        ]
      },
      "UpdateMenuDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "菜单名称",
            "example": "用户管理"
          },
          "path": {
            "type": "string",
            "description": "菜单路径",
            "example": "/users"
          },
          "acl": {
            "type": "string",
            "description": "权限策略",
            "example": "user:read,user:write"
          },
          "roleId": {
            "type": "number",
            "description": "角色ID",
            "example": 1
          }
        }
      }
    }
  }
}